@@ -78,56 +73,57 @@
   },
   "homepage": "https://github.com/4Catalyzer/found#readme",
   "dependencies": {
-    "@babel/runtime-corejs2": "^7.4.0",
+    "@babel/runtime-corejs3": "^7.6.2",
+    "@restart/context": "^2.1.4",
     "farce": "^0.2.8",
+    "invariant": "^2.2.4",
     "is-promise": "^2.1.0",
-    "lodash": "^4.17.11",
+    "lodash": "^4.17.15",
     "path-to-regexp": "^1.7.0",
     "prop-types": "^15.7.2",
-    "react-redux": "^6.0.1",
+    "react-redux": "^7.0.0",
     "react-static-container": "^1.0.2",
-    "redux": "^4.0.1",
+    "redux": "^4.0.4",
     "warning": "^4.0.3"
   },
   "peerDependencies": {
-    "react": ">=16.4.0"
+    "react": ">=16.8.0"
   },
   "devDependencies": {
-    "@4c/babel-preset": "^5.0.4",
-    "@babel/cli": "^7.2.3",
-    "@babel/core": "^7.4.0",
-    "@types/react": "^16.8.8",
-    "@types/react-dom": "^16.8.2",
-    "babel-eslint": "^10.0.1",
-    "babel-jest": "^24.5.0",
-    "babel-plugin-add-module-exports": "^1.0.0",
-    "codecov": "^3.2.0",
-    "cpy-cli": "^2.0.0",
-    "delay": "^4.1.0",
+    "@4c/babel-preset": "^7.2.4",
+    "@babel/cli": "^7.7.7",
+    "@babel/core": "^7.7.7",
+    "@types/react": "^16.9.17",
+    "@types/react-dom": "^16.9.4",
+    "babel-jest": "^24.9.0",
+    "babel-plugin-add-module-exports": "^1.0.2",
+    "codecov": "^3.6.1",
+    "cpy-cli": "^3.0.0",
+    "delay": "^4.3.0",
     "doctoc": "^1.4.0",
-    "dtslint": "^0.5.5",
-    "enzyme": "^3.9.0",
-    "enzyme-adapter-react-16": "^1.11.2",
-    "eslint": "^5.15.3",
-    "eslint-config-4catalyzer-jest": "^1.2.1",
-    "eslint-config-4catalyzer-react": "^0.9.3",
-    "eslint-config-prettier": "^4.1.0",
-    "eslint-plugin-import": "^2.16.0",
-    "eslint-plugin-jest": "^22.4.1",
-    "eslint-plugin-jsx-a11y": "^6.2.1",
-    "eslint-plugin-prettier": "^3.0.1",
-    "eslint-plugin-react": "^7.12.4",
-    "husky": "^1.3.1",
-    "jest": "^24.5.0",
-    "lint-staged": "^8.1.5",
-    "p-defer": "^2.0.0",
-    "prettier": "^1.16.4",
-    "react": "^16.8.4",
-    "react-dom": "^16.8.4",
+    "dtslint": "^2.0.3",
+    "enzyme": "^3.11.0",
+    "enzyme-adapter-react-16": "^1.15.2",
+    "eslint": "^6.8.0",
+    "eslint-config-4catalyzer-jest": "^2.0.2",
+    "eslint-config-4catalyzer-react": "^1.0.1",
+    "eslint-config-prettier": "^6.9.0",
+    "eslint-plugin-import": "^2.19.1",
+    "eslint-plugin-jest": "^23.3.0",
+    "eslint-plugin-jsx-a11y": "^6.2.3",
+    "eslint-plugin-prettier": "^3.1.2",
+    "eslint-plugin-react": "^7.17.0",
+    "eslint-plugin-react-hooks": "^2.3.0",
+    "husky": "^4.0.0",
+    "jest": "^24.9.0",
+    "lint-staged": "^9.5.0",
+    "p-defer": "^3.0.0",
+    "prettier": "^1.19.1",
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0",
     "react-proxy": "^3.0.0-alpha.1",
-    "react-stand-in": "^4.0.0-beta.14",
-    "react-test-renderer": "^16.8.4",
-    "rimraf": "^2.6.3",
-    "typescript": "^3.3.4000"
+    "react-stand-in": "^4.0.0-beta.21",
+    "react-test-renderer": "^16.12.0",
+    "rimraf": "^3.0.0"
   }
 }
