@@ -1,36 +1,34 @@
 import isEqual from 'lodash/isEqual';
 import PropTypes from 'prop-types';
 import React from 'react';
+import { ReactReduxContext } from 'react-redux';
 import StaticContainer from 'react-static-container';
 import warning from 'warning';
+import mapContextToProps from '@restart/context/mapContextToProps';
 
-import { routerShape } from './PropTypes';
 import createRender from './createRender';
+import RouterContext from './RouterContext';
+import createStoreRouterObject from './utils/createStoreRouterObject';
 import resolveRenderArgs from './utils/resolveRenderArgs';
 
 export default function createBaseRouter({
-  render,
   renderPending,
   renderReady,
   renderError,
+  render = createRender({
+    renderPending,
+    renderReady,
+    renderError,
+  }),
 }) {
-  // eslint-disable-next-line no-param-reassign
-  render =
-    render ||
-    createRender({
-      renderPending,
-      renderReady,
-      renderError,
-    });
-
   const propTypes = {
+    store: PropTypes.object.isRequired,
     match: PropTypes.object.isRequired,
     resolvedMatch: PropTypes.object.isRequired,
     matchContext: PropTypes.any,
     resolver: PropTypes.shape({
       resolveElements: PropTypes.func.isRequired,
     }).isRequired,
-    router: routerShape.isRequired,
     onResolveMatch: PropTypes.func.isRequired,
     initialRenderArgs: PropTypes.object,
   };
@@ -39,16 +37,30 @@ export default function createBaseRouter({
     constructor(props) {
       super(props);
 
-      const { initialRenderArgs } = props;
+      const {
+        store,
+        match,
+        resolver,
+        matchContext,
+        initialRenderArgs,
+      } = props;
+
+      this.router = createStoreRouterObject(store);
 
       this.state = {
+        isInitialRender: true,
+        match,
+        resolver,
+        matchContext,
+        iteration: 0,
+        renderArgs: initialRenderArgs || null,
         element: initialRenderArgs ? render(initialRenderArgs) : null,
       };
 
       this.mounted = true;
       this.lastRenderArgs = initialRenderArgs;
 
-      this.shouldResolveMatch = false;
+      this.lastIteration = 0;
       this.pendingResolvedMatch = false;
     }
 
@@ -68,35 +80,40 @@ export default function createBaseRouter({
         if (window.__FOUND_HOT_RELOAD__) {
           warning(
             !window.__FOUND_REPLACE_ROUTE_CONFIG__,
-            'Replacing existing hot reloading hook. Do not render more than ' +
-              'one router instance when using hot reloading.',
+            'Replacing existing hot reloading hook. Do not render more than one router instance when using hot reloading.',
           );
 
-          window.__FOUND_REPLACE_ROUTE_CONFIG__ = this.props.router.replaceRouteConfig;
+          window.__FOUND_REPLACE_ROUTE_CONFIG__ = this.router.replaceRouteConfig;
         }
         /* eslint-enable no-underscore-dangle */
         /* eslint-env browser: false */
       }
     }
 
-    componentWillReceiveProps(nextProps) {
-      warning(
-        nextProps.router === this.props.router,
-        '<BaseRouter> does not support changing the router object.',
-      );
+    static getDerivedStateFromProps({ match, resolver, matchContext }, state) {
+      if (state.isInitialRender) {
+        return { isInitialRender: false };
+      }
 
       if (
-        nextProps.match !== this.props.match ||
-        nextProps.resolver !== this.props.resolver ||
-        !isEqual(nextProps.matchContext, this.props.matchContext)
+        match !== state.match ||
+        resolver !== state.resolver ||
+        !isEqual(matchContext, state.matchContext)
       ) {
-        this.shouldResolveMatch = true;
+        return {
+          match,
+          resolver,
+          matchContext,
+          iteration: state.iteration + 1,
+        };
       }
+
+      return null;
     }
 
     componentDidUpdate() {
-      if (this.shouldResolveMatch) {
-        this.shouldResolveMatch = false;
+      if (this.state.iteration > this.lastIteration) {
+        this.lastIteration = this.state.iteration;
         this.resolveMatch();
       }
     }
@@ -115,12 +132,23 @@ export default function createBaseRouter({
       }
     }
 
+    componentDidCatch(error) {
+      if (this.handleMaybeRedirectException(error)) {
+        return;
+      }
+
+      if (!error.isFoundHttpError) {
+        this.lastRenderArgs = { ...this.lastRenderArgs, error };
+        this.updateElement();
+      }
+    }
+
     handleMaybeRedirectException(e) {
       if (!e.isFoundRedirectException) {
         return false;
       }
 
-      this.props.router.replace(e.location);
+      this.router.replace(e.location);
       return true;
     }
 
@@ -128,7 +156,11 @@ export default function createBaseRouter({
       const pendingMatch = this.props.match;
 
       try {
-        for await (const renderArgs of resolveRenderArgs(this.props)) {
+        for await (const renderArgs of resolveRenderArgs(
+          this.router,
+          this.props,
+        )) {
+          // Don't do anything if we're resolving an outdated match.
           if (!this.mounted || this.props.match !== pendingMatch) {
             return;
           }
@@ -152,6 +184,10 @@ export default function createBaseRouter({
           }
         }
       } catch (e) {
+        if (!this.mounted || this.props.match !== pendingMatch) {
+          return;
+        }
+
         if (this.handleMaybeRedirectException(e)) {
           return;
         }
@@ -161,31 +197,34 @@ export default function createBaseRouter({
     }
 
     updateElement() {
+      const renderArgs = this.lastRenderArgs;
+
       this.setState({
-        element: render(this.lastRenderArgs),
+        renderArgs,
+        element: render(renderArgs),
       });
     }
 
-    componentDidCatch(error) {
-      if (this.handleMaybeRedirectException(error)) {
-        return;
-      }
-
-      if (error.isFoundHttpError) {
-        this.lastRenderArgs = { ...this.lastRenderArgs, error };
-        this.updateElement();
-      }
-    }
-
     render() {
+      const { iteration, renderArgs, element } = this.state;
+
       // Don't rerender synchronously if we have another rerender coming. Just
       // memoizing the element here doesn't do anything because we're using
       // context.
       return (
         <StaticContainer
-          shouldUpdate={!this.shouldResolveMatch && !this.pendingResolvedMatch}
+          shouldUpdate={
+            this.lastIteration === iteration && !this.pendingResolvedMatch
+          }
         >
-          {this.state.element}
+          <RouterContext.Provider
+            value={{
+              router: this.router,
+              match: renderArgs,
+            }}
+          >
+            {element}
+          </RouterContext.Provider>
         </StaticContainer>
       );
     }
@@ -193,5 +232,13 @@ export default function createBaseRouter({
 
   BaseRouter.propTypes = propTypes;
 
-  return BaseRouter;
+  // FIXME: For some reason, using contextType doesn't work here.
+  return mapContextToProps(
+    {
+      consumers: ReactReduxContext,
+      mapToProps: ({ store }) => ({ store }),
+      displayName: 'withStore(BaseRouter)',
+    },
+    BaseRouter,
+  );
 }
