@@ -52,10 +52,7 @@ const BrowserRouter = createBrowserRouter({
         },
         {
           path: 'bar',
-          getComponent: () =>
-            new Promise(resolve => {
-              setTimeout(resolve, 1000, ({ data }) => <div>{data}</div>);
-            }),
+          getComponent: () => import('./Bar').then(m => m.default),
           getData: () =>
             new Promise(resolve => {
               setTimeout(resolve, 1000, 'Bar');
@@ -79,9 +76,11 @@ const BrowserRouter = createBrowserRouter({
     },
   ],
 
-  renderError: (
-    { error }, // eslint-disable-line react/prop-types
-  ) => <div>{error.status === 404 ? 'Not found' : 'Error'}</div>,
+  /* eslint-disable react/prop-types */
+  renderError: ({ error }) => (
+    <div>{error.status === 404 ? 'Not found' : 'Error'}</div>
+  ),
+  /* eslint-enable react/prop-types */
 });
 
 ReactDOM.render(<BrowserRouter />, document.getElementById('root'));
