@@ -7,9 +7,11 @@ import routeConfig from './routeConfig';
 const BrowserRouter = createBrowserRouter({
   routeConfig,
 
-  renderError: (
-    { error }, // eslint-disable-line react/prop-types
-  ) => <div>{error.status === 404 ? 'Not found' : 'Error'}</div>,
+  /* eslint-disable react/prop-types */
+  renderError: ({ error }) => (
+    <div>{error.status === 404 ? 'Not found' : 'Error'}</div>
+  ),
+  /* eslint-enable react/prop-types */
 });
 
 export default hot(module)(BrowserRouter);
