@@ -106,9 +106,11 @@ store.dispatch(FarceActions.init());
 
 const ConnectedRouter = createConnectedRouter({
   render: createRender({
-    renderError: (
-      { error }, // eslint-disable-line react/prop-types
-    ) => <div>{error.status === 404 ? 'Not found' : 'Error'}</div>,
+    /* eslint-disable react/prop-types */
+    renderError: ({ error }) => (
+      <div>{error.status === 404 ? 'Not found' : 'Error'}</div>
+    ),
+    /* eslint-enable react/prop-types */
   }),
 });
 
