@@ -2,7 +2,9 @@ import createRender from 'found/lib/createRender';
 import React from 'react';
 
 export default createRender({
-  renderError: (
-    { error }, // eslint-disable-line react/prop-types
-  ) => <div>{error.status === 404 ? 'Not found' : 'Error'}</div>,
+  /* eslint-disable react/prop-types */
+  renderError: ({ error }) => (
+    <div>{error.status === 404 ? 'Not found' : 'Error'}</div>
+  ),
+  /* eslint-enable react/prop-types */
 });
