@@ -2,7 +2,6 @@ export ActionTypes from './ActionTypes';
 export BaseLink from './BaseLink';
 export createBaseRouter from './createBaseRouter';
 export createBrowserRouter from './createBrowserRouter';
-export createConnectedLink from './createConnectedLink';
 export createConnectedRouter from './createConnectedRouter';
 export createElements from './createElements';
 export createFarceRouter from './createFarceRouter';
@@ -10,7 +9,6 @@ export createInitialBrowserRouter from './createInitialBrowserRouter';
 export createInitialFarceRouter from './createInitialFarceRouter';
 export createMatchEnhancer from './createMatchEnhancer';
 export createRender from './createRender';
-export createWithRouter from './createWithRouter';
 export ElementsRenderer from './ElementsRenderer';
 export foundReducer from './foundReducer';
 export getRenderArgs from './getRenderArgs';
@@ -26,4 +24,6 @@ export RedirectException from './RedirectException';
 export resolver from './resolver';
 export * as ResolverUtils from './ResolverUtils';
 export Route from './Route';
+export RouterContext from './RouterContext';
+export useRouter from './useRouter';
 export withRouter from './withRouter';
