@@ -1,3 +1,14 @@
-import createWithRouter from './createWithRouter';
+import mapContextToProps from '@restart/context/mapContextToProps';
 
-export default createWithRouter({});
+import RouterContext from './RouterContext';
+
+export default function withRouter(Component) {
+  return mapContextToProps(
+    {
+      consumers: RouterContext,
+      mapToProps: context => context,
+      displayName: `withRouter(${Component.displayName || Component.name})`,
+    },
+    Component,
+  );
+}
