@@ -87,13 +87,22 @@ describe('<BaseLink>', () => {
       expect(router.push).not.toBeCalled();
     });
 
-    it('should not navigate if target is defined', () => {
+    it('should not navigate if target is defined and not _self', () => {
       const link = mount(
         <BaseLink to="/" match={{}} router={router} target="_blank" />,
       );
 
       link.find('a').simulate('click', { button: 0 });
       expect(router.push).not.toBeCalled();
+    });
+
+    it('should navigate if target is _self', () => {
+      const link = mount(
+        <BaseLink to="/" match={{}} router={router} target="_self" />,
+      );
+
+      link.find('a').simulate('click', { button: 0 });
+      expect(router.push).toBeCalled();
     });
   });
 
